requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
PyMuPDF==1.23.7
python-docx==0.8.11
docx2txt==0.8
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
email-validator==2.1.0.post1 